Move original function behind wall, add proxy
